Some styling done
